requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Admin==1.6.1
Flask-WTF==1.2.1
WTForms==3.1.1
email-validator==2.1.0
Werkzeug==3.0.1
python-dotenv==1.0.0

# Payment Processing
stripe==8.1.0
paypalrestsdk==1.7.1

# Email & Notifications
Flask-Mail==0.9.1
python-socketio==5.9.0
python-engineio==4.8.0

# Real-time Communication
Flask-SocketIO==5.3.5
python-socketio==5.9.0

# Scheduling & Tasks
celery==5.3.4
redis==5.0.1
APScheduler==3.10.4

# Maps & Geolocation
geopy==2.4.0

# PDF & Document Generation
reportlab==4.0.9
python-docx==0.8.11
PyPDF2==3.0.1

# AI/ML for Recommendations
scikit-learn==1.4.0
pandas==2.1.4

# Analytics
plotly==5.18.0

# Security & Hashing
PyJWT==2.8.1
requests==2.31.0

# Data Validation
pydantic==2.5.3

# Image Processing
Pillow==10.1.0
